Fixed accuracy of energy reported: no more floats but Wh instead.

+++ lib/MycilaRouter/MycilaRouter.h
@@ -20,7 +20,7 @@ namespace Mycila {
       typedef struct {
           float apparentPower = 0;
           float current = 0;
-          float energy = 0;
+          uint32_t energy = 0;
           float power = 0;
           float powerFactor = NAN;
           float resistance = NAN;
